+++ src/main/kotlin/mobi/hsz/idea/gitignore/actions/CreateUserTemplateAction.kt
@@ -11,6 +11,7 @@ import mobi.hsz.idea.gitignore.IgnoreBundle
 import mobi.hsz.idea.gitignore.psi.IgnoreFile
 import mobi.hsz.idea.gitignore.ui.template.UserTemplateDialog
 import mobi.hsz.idea.gitignore.util.Icons
+import com.intellij.openapi.vcs.changes.ignore.psi.IgnoreFile as NativeIgnoreFile
 
 /**
  * Action that creates new user template with predefined content - i.e. from currently opened file.
@@ -24,11 +25,11 @@ class CreateUserTemplateAction : AnAction(
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.getData(CommonDataKeys.PROJECT) ?: return
         val file = e.getData(CommonDataKeys.PSI_FILE)
-        if (file !is IgnoreFile) {
+        if (file !is IgnoreFile && file !is NativeIgnoreFile) {
             return
         }
-        var content = file.getText()
-        file.getViewProvider().document?.let { document ->
+        var content = file.text
+        file.viewProvider.document?.let { document ->
             EditorFactory.getInstance().getEditors(document).first()?.let { editor ->
                 val selectedText = editor.selectionModel.selectedText
                 if (!StringUtil.isEmpty(selectedText)) {
@@ -40,12 +41,19 @@ class CreateUserTemplateAction : AnAction(
     }
 
     override fun update(e: AnActionEvent) {
-        val file = e.getData(CommonDataKeys.PSI_FILE)
-        if (file !is IgnoreFile) {
-            e.presentation.isVisible = false
-            return
+        when (val file = e.getData(CommonDataKeys.PSI_FILE)) {
+            is IgnoreFile -> {
+                e.presentation.icon = file.fileType.icon
+            }
+
+            is NativeIgnoreFile -> {
+                e.presentation.icon = Icons.GIT
+            }
+
+            else -> {
+                e.presentation.isVisible = false
+            }
         }
-        templatePresentation.icon = file.fileType.icon
     }
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT